src/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.121.2",
    "langchain>=1.0.7",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.0.3",
    "neo4j>=6.0.3",
    "openai>=2.8.0",
    "python-dotenv>=1.2.1",
    "tiktoken>=0.12.0",
]
